feat(events): Embed add-on definition CRUD in event create/edit pages
- Extract shared AddOnDefinitionEditor component with dual-mode support:
  - Live mode (edit page): CRUD via API hooks when eventId exists
  - Local mode (create page): definitions queued in React state, created post-save
- Embed AddOnDefinitionEditor inside AddOnConfigForm (replaces guidance banner)
- EventCreationForm: add pendingAddOnDefinitions state, create definitions
  via Promise.all after event creation
- EventEditForm: pass eventId to AddOnConfigForm for live-mode editing
- AddOnsManagementTab: replace ~250 lines of inline CRUD with shared editor
- Remove "Next step: go to manage page" guidance (no longer needed)
